+++ opencascade-hs/src/OpenCascade/BRepGProp.hs
@@ -1,5 +1,9 @@
 {-# LANGUAGE CApiFFI #-}
-module OpenCascade.BRepGProp (volumeProperties) where
+module OpenCascade.BRepGProp 
+( volumeProperties
+, surfaceProperties
+, linearProperties
+) where
 
 import OpenCascade.TopoDS.Types (Shape)
 import OpenCascade.GProp.Types (GProps)
@@ -12,4 +16,18 @@ foreign import capi unsafe "hs_BRepGProp.h hs_BRepGProp_VolumeProperties" rawVol
 volumeProperties :: Ptr Shape -> Ptr GProps -> Bool -> Bool -> Bool -> IO ()
 volumeProperties shape props onlyClosed skipShared useTriangulation =
     rawVolumeProperties shape props (boolToCBool onlyClosed) (boolToCBool skipShared) (boolToCBool useTriangulation)
+
+foreign import capi unsafe "hs_BRepGProp.h hs_BRepGProp_SurfaceProperties" rawSurfaceProperties :: Ptr Shape -> Ptr GProps -> CBool -> CBool -> IO ()
+
+surfaceProperties :: Ptr Shape -> Ptr GProps -> Bool -> Bool -> IO ()
+surfaceProperties shape props skipShared useTriangulation =
+    rawSurfaceProperties shape props (boolToCBool skipShared) (boolToCBool useTriangulation)
+
+    
+foreign import capi unsafe "hs_BRepGProp.h hs_BRepGProp_LinearProperties" rawLinearProperties :: Ptr Shape -> Ptr GProps -> CBool -> CBool -> IO ()
+
+linearProperties :: Ptr Shape -> Ptr GProps -> Bool -> Bool -> IO ()
+linearProperties shape props skipShared useTriangulation =
+    rawLinearProperties shape props (boolToCBool skipShared) (boolToCBool useTriangulation)
+